Gene Segment Start Stop Finder
This code will help automate part of the gene annotation workflow by eliminating the need to manually verify a gene sequence's position in a larger gene sequence. Given two FASTA files, the program will correctly output the start and stop nucleotide positions of a given gene segment in the form of a comma-separated value (.csv) file.  
To run the Python script, please make sure that you have a proper Python environment installed on your computer with the Biopython library installed. Personally, I use the Anaconda navigator to handle all of my libraries because it is open-source and free. 
Please do not hesitate to contact me if you have any questions or concerns.